auth refactored..successfully!
